refactor: move chunk helper to src/helpers/arrays.ts
