+++ src/helpers/arrays.ts
@@ -1,0 +1,10 @@
+export function chunk<T>(array: T[], size: number): T[][] {
+  if (size <= 0) {
+    throw new Error(`chunk size must be positive, got ${size}`);
+  }
+  const chunks: T[][] = [];
+  for (let i = 0; i < array.length; i += size) {
+    chunks.push(array.slice(i, i + size));
+  }
+  return chunks;
+}
